MODIFY : modified price, discount_rate data type

+++ product/views.py
@@ -22,15 +22,15 @@ class CategoryView(View):
 
             products_list = list()
             for product in products_obj_list:
-                discount_rate    = int(DiscountRate.objects.get(product=product).rate * 100)
-                discounted_price = int(product.price - product.price * (discount_rate / 100))
+                discount_rate    = DiscountRate.objects.get(product=product).rate * 100
+                discounted_price = product.price - product.price * (discount_rate / 100)
 
                 product_dict = dict(
                                     product_name      = product.name,
                                     product_price     = int(product.price),
                                     product_thumbnail = product.thumbnail_image_url,
-                                    discount_rate     = discount_rate,
-                                    discounted_price  = discounted_price
+                                    discount_rate     = int(discount_rate),
+                                    discounted_price  = int(discounted_price)
                                 )
                 products_list.append(product_dict)
             return JsonResponse({'results': products_list}, status=200)
